fix symbol versions
versions need to include the epoch. symbols files are eventually
put into the final tarball allowing a linkee to decide which version
of the library is required to fullfill the linkees symbol dependencies.
this will use the version association of a symbol to decide which version
this should be, so if our versions do not include an epoch the version
relationship between linkee and library are wrong (i.e. missing the
epoch)