bring back xvfb-run wrapping
testWindowmanagement was recently changed to use QTEST_MAIN, which
instantiates a QGuiApplication which will load the xcb plugin, which
will abort if X is not present.
Checked back with mgraesslin and this change was intentional, so we need
the xvfb wrapping again to provision an X server for that test.